Rename module alloc to avoid name conflict
